requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.1
dnspython==1.16.0
email-validator==1.0.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.1
pycparser==2.20
python-dotenv==0.13.0
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.3.0
